# testbench/json-api/src/openapi.yaml
openapi: 3.0.3
info:
  title: Argus Testbench JSON API
  description: >
    A simple CRUD REST API for testing Argus's API discovery feature.
    Provides users, posts, comments, and todos resources with pagination,
    filtering, and full CRUD operations. Data is stored in-memory and
    seeded with sample records on startup.
  version: 1.0.0
  contact:
    name: Argus Testbench
servers:
  - url: "{baseUrl}/api/v1"
    variables:
      baseUrl:
        default: "http://localhost:3000"
        description: Base URL of the JSON API server

paths:
  /users:
    get:
      tags: [Users]
      summary: List users
      description: Returns a paginated list of users.
      operationId: listUsers
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated list of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
    post:
      tags: [Users]
      summary: Create a user
      description: Creates a new user with the given name and email.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /users/{id}:
    get:
      tags: [Users]
      summary: Get a user by ID
      operationId: getUser
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags: [Users]
      summary: Update a user
      operationId: updateUser
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Users]
      summary: Delete a user
      operationId: deleteUser
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "204":
          description: User deleted
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /posts:
    get:
      tags: [Posts]
      summary: List posts
      description: Returns a paginated list of posts. Optionally filter by userId.
      operationId: listPosts
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: userId
          in: query
          description: Filter posts by user ID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Paginated list of posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Post"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
    post:
      tags: [Posts]
      summary: Create a post
      operationId: createPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /posts/{id}:
    get:
      tags: [Posts]
      summary: Get a post by ID
      operationId: getPost
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Post found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags: [Posts]
      summary: Update a post
      operationId: updatePost
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostRequest"
      responses:
        "200":
          description: Post updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Posts]
      summary: Delete a post
      operationId: deletePost
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "204":
          description: Post deleted
        "404":
          description: Post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /posts/{id}/comments:
    get:
      tags: [Posts]
      summary: Get comments for a post
      operationId: getPostComments
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: List of comments for the post
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Comment"
        "404":
          description: Post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /comments:
    get:
      tags: [Comments]
      summary: List comments
      description: Returns a paginated list of comments. Optionally filter by postId.
      operationId: listComments
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: postId
          in: query
          description: Filter comments by post ID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Paginated list of comments
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Comment"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
    post:
      tags: [Comments]
      summary: Create a comment
      operationId: createComment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /comments/{id}:
    get:
      tags: [Comments]
      summary: Get a comment by ID
      operationId: getComment
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Comment found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "404":
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags: [Comments]
      summary: Update a comment
      operationId: updateComment
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCommentRequest"
      responses:
        "200":
          description: Comment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "404":
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Comments]
      summary: Delete a comment
      operationId: deleteComment
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "204":
          description: Comment deleted
        "404":
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /todos:
    get:
      tags: [Todos]
      summary: List todos
      description: Returns a paginated list of todos. Optionally filter by userId and/or completed status.
      operationId: listTodos
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: userId
          in: query
          description: Filter todos by user ID
          schema:
            type: string
            format: uuid
        - name: completed
          in: query
          description: Filter by completion status
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        "200":
          description: Paginated list of todos
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Todo"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
    post:
      tags: [Todos]
      summary: Create a todo
      operationId: createTodo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTodoRequest"
      responses:
        "201":
          description: Todo created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /todos/{id}:
    get:
      tags: [Todos]
      summary: Get a todo by ID
      operationId: getTodo
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Todo found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "404":
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      tags: [Todos]
      summary: Partially update a todo
      description: Update specific fields of a todo (e.g., toggle completed, change title).
      operationId: patchTodo
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchTodoRequest"
      responses:
        "200":
          description: Todo updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "404":
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Todos]
      summary: Delete a todo
      operationId: deleteTodo
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "204":
          description: Todo deleted
        "404":
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      description: Resource UUID
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      description: Items per page (max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10

  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message

    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time

    CreateUserRequest:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          example: "Jane Doe"
        email:
          type: string
          format: email
          example: "jane@example.com"

    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email

    Post:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        body:
          type: string
        userId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time

    CreatePostRequest:
      type: object
      required: [title, body, userId]
      properties:
        title:
          type: string
          example: "My First Post"
        body:
          type: string
          example: "This is the content of the post."
        userId:
          type: string
          format: uuid

    UpdatePostRequest:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        userId:
          type: string
          format: uuid

    Comment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        body:
          type: string
        postId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time

    CreateCommentRequest:
      type: object
      required: [body, postId, userId]
      properties:
        body:
          type: string
          example: "Great post!"
        postId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid

    UpdateCommentRequest:
      type: object
      properties:
        body:
          type: string
        postId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid

    Todo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        userId:
          type: string
          format: uuid
        completed:
          type: boolean
        createdAt:
          type: string
          format: date-time

    CreateTodoRequest:
      type: object
      required: [title, userId]
      properties:
        title:
          type: string
          example: "Buy groceries"
        userId:
          type: string
          format: uuid
        completed:
          type: boolean
          default: false

    PatchTodoRequest:
      type: object
      properties:
        title:
          type: string
        completed:
          type: boolean